remove useless code.
