+++ executor/builder.go
@@ -14,26 +14,16 @@
 package executor
 
 import (
-	"math"
-	"strings"
-
 	"github.com/juju/errors"
-	"github.com/ngaut/log"
 	"github.com/pingcap/tidb/ast"
 	"github.com/pingcap/tidb/context"
 	"github.com/pingcap/tidb/expression"
 	"github.com/pingcap/tidb/infoschema"
-	"github.com/pingcap/tidb/kv"
 	"github.com/pingcap/tidb/model"
-	"github.com/pingcap/tidb/mysql"
 	"github.com/pingcap/tidb/parser/opcode"
 	"github.com/pingcap/tidb/plan"
 	"github.com/pingcap/tidb/sessionctx/autocommit"
 	"github.com/pingcap/tidb/sessionctx/variable"
-	"github.com/pingcap/tidb/table"
-	"github.com/pingcap/tidb/util/charset"
-	"github.com/pingcap/tidb/util/types"
-	"github.com/pingcap/tipb/go-tipb"
 )
 
 // executorBuilder builds an Executor from a Plan.
@@ -55,16 +45,12 @@ func (b *executorBuilder) build(p plan.Plan) Executor {
 	switch v := p.(type) {
 	case nil:
 		return nil
-	case *plan.Aggregate:
-		return b.buildAggregate(v)
 	case *plan.CheckTable:
 		return b.buildCheckTable(v)
 	case *plan.DDL:
 		return b.buildDDL(v)
 	case *plan.Deallocate:
 		return b.buildDeallocate(v)
-	case *plan.Delete:
-		return b.buildDelete(v)
 	case *plan.NewDelete:
 		return b.buildNewDelete(v)
 	case *plan.Distinct:
@@ -76,22 +62,12 @@ func (b *executorBuilder) build(p plan.Plan) Executor {
 	case *plan.Filter:
 		src := b.build(v.GetChildByIndex(0))
 		return b.buildFilter(src, v.Conditions)
-	case *plan.Having:
-		return b.buildHaving(v)
-	case *plan.IndexScan:
-		return b.buildIndexScan(v)
 	case *plan.Insert:
 		return b.buildInsert(v)
-	case *plan.JoinInner:
-		return b.buildJoinInner(v)
-	case *plan.JoinOuter:
-		return b.buildJoinOuter(v)
 	case *plan.Limit:
 		return b.buildLimit(v)
 	case *plan.Prepare:
 		return b.buildPrepare(v)
-	case *plan.SelectFields:
-		return b.buildSelectFields(v)
 	case *plan.SelectLock:
 		return b.buildSelectLock(v)
 	case *plan.ShowDDL:
@@ -100,20 +76,10 @@ func (b *executorBuilder) build(p plan.Plan) Executor {
 		return b.buildShow(v)
 	case *plan.Simple:
 		return b.buildSimple(v)
-	case *plan.Sort:
-		return b.buildSort(v)
 	case *plan.NewSort:
 		return b.buildNewSort(v)
-	case *plan.TableDual:
-		return b.buildTableDual(v)
-	case *plan.TableScan:
-		return b.buildTableScan(v)
-	case *plan.Union:
-		return b.buildUnion(v)
 	case *plan.NewUnion:
 		return b.buildNewUnion(v)
-	case *plan.Update:
-		return b.buildUpdate(v)
 	case *plan.NewUpdate:
 		return b.buildNewUpdate(v)
 	case *plan.PhysicalHashJoin:
@@ -159,64 +125,6 @@ func (b *executorBuilder) buildFilter(src Executor, conditions []ast.ExprNode) E
 	}
 }
 
-func (b *executorBuilder) buildTableDual(v *plan.TableDual) Executor {
-	e := &TableDualExec{fields: v.Fields()}
-	return b.buildFilter(e, v.FilterConditions)
-}
-
-func (b *executorBuilder) buildTableScan(v *plan.TableScan) Executor {
-	txn, err := b.ctx.GetTxn(false)
-	if err != nil {
-		b.err = err
-		return nil
-	}
-	table, _ := b.is.TableByID(v.Table.ID)
-	client := txn.GetClient()
-	var memDB bool
-	switch v.Fields()[0].DBName.L {
-	case "information_schema", "performance_schema":
-		memDB = true
-	}
-	supportDesc := client.SupportRequestType(kv.ReqTypeSelect, kv.ReqSubTypeDesc)
-	if !memDB && client.SupportRequestType(kv.ReqTypeSelect, 0) {
-		log.Debug("xapi select table")
-		e := &XSelectTableExec{
-			table:       table,
-			ctx:         b.ctx,
-			tablePlan:   v,
-			supportDesc: supportDesc,
-		}
-		where, remained := b.conditionsToPBExpr(client, v.FilterConditions, v.TableName)
-		if where != nil {
-			e.where = where
-		}
-		if len(remained) == 0 {
-			e.allFiltersPushed = true
-		}
-		var ex Executor
-		if txn.IsReadOnly() {
-			ex = e
-		} else {
-			ex = b.buildUnionScanExec(e)
-		}
-		return b.buildFilter(ex, remained)
-	}
-
-	e := &TableScanExec{
-		t:           table,
-		tableAsName: v.TableAsName,
-		fields:      v.Fields(),
-		ctx:         b.ctx,
-		ranges:      v.Ranges,
-		seekHandle:  math.MinInt64,
-	}
-	x := b.buildFilter(e, v.FilterConditions)
-	if v.Desc {
-		x = &ReverseExec{Src: x}
-	}
-	return x
-}
-
 func (b *executorBuilder) buildShowDDL(v *plan.ShowDDL) Executor {
 	return &ShowDDLExec{
 		fields: v.Fields(),
@@ -236,118 +144,6 @@ func (b *executorBuilder) buildDeallocate(v *plan.Deallocate) Executor {
 		ctx:  b.ctx,
 		Name: v.Name,
 	}
-}
-
-func (b *executorBuilder) buildIndexScan(v *plan.IndexScan) Executor {
-	txn, err := b.ctx.GetTxn(false)
-	if err != nil {
-		b.err = err
-		return nil
-	}
-	tbl, _ := b.is.TableByID(v.Table.ID)
-
-	client := txn.GetClient()
-	supportDesc := client.SupportRequestType(kv.ReqTypeIndex, kv.ReqSubTypeDesc)
-	var memDB bool
-	switch v.Fields()[0].DBName.L {
-	case "information_schema", "performance_schema":
-		memDB = true
-	}
-	if !memDB && client.SupportRequestType(kv.ReqTypeIndex, 0) {
-		log.Debug("xapi select index")
-		e := &XSelectIndexExec{
-			table:       tbl,
-			ctx:         b.ctx,
-			indexPlan:   v,
-			supportDesc: supportDesc,
-		}
-		where, remained := b.conditionsToPBExpr(client, v.FilterConditions, v.TableName)
-		if where != nil {
-			e.where = where
-		}
-		var ex Executor
-		if txn.IsReadOnly() {
-			ex = e
-		} else {
-			ex = b.buildUnionScanExec(e)
-		}
-		if v.Index.HasPrefixIndex() {
-			ex = b.buildFilter(ex, v.AccessConditions)
-		}
-		return b.buildFilter(ex, remained)
-	}
-
-	var idx table.Index
-	for _, val := range tbl.Indices() {
-		if val.Meta().Name.L == v.Index.Name.L {
-			idx = val
-			break
-		}
-	}
-
-	valueTypes := make([]*types.FieldType, len(idx.Meta().Columns))
-	for i, ic := range idx.Meta().Columns {
-		col := tbl.Cols()[ic.Offset]
-		valueTypes[i] = &col.FieldType
-	}
-
-	e := &IndexScanExec{
-		tbl:         tbl,
-		tableAsName: v.TableAsName,
-		idx:         idx,
-		fields:      v.Fields(),
-		ctx:         b.ctx,
-		Desc:        v.Desc,
-		valueTypes:  valueTypes,
-	}
-
-	e.Ranges = make([]*IndexRangeExec, len(v.Ranges))
-	for i, val := range v.Ranges {
-		e.Ranges[i] = b.buildIndexRange(e, val, idx.Meta().Columns)
-	}
-	x := Executor(e)
-	if v.Index.HasPrefixIndex() {
-		x = b.buildFilter(x, append(v.AccessConditions, v.FilterConditions...))
-	} else {
-		x = b.buildFilter(x, v.FilterConditions)
-	}
-	if v.Desc {
-		x = &ReverseExec{Src: x}
-	}
-	return x
-}
-
-func (b *executorBuilder) buildIndexRange(scan *IndexScanExec, v *plan.IndexRange, ics []*model.IndexColumn) *IndexRangeExec {
-	ran := &IndexRangeExec{
-		scan:        scan,
-		lowVals:     v.LowVal,
-		lowExclude:  v.LowExclude,
-		highVals:    v.HighVal,
-		highExclude: v.HighExclude,
-	}
-	return ran
-}
-
-func (b *executorBuilder) buildJoinOuter(v *plan.JoinOuter) *JoinOuterExec {
-	e := &JoinOuterExec{
-		OuterExec: b.build(v.Outer),
-		InnerPlan: v.Inner,
-		fields:    v.Fields(),
-		builder:   b,
-	}
-	return e
-}
-
-func (b *executorBuilder) buildJoinInner(v *plan.JoinInner) *JoinInnerExec {
-	e := &JoinInnerExec{
-		InnerPlans: v.Inners,
-		innerExecs: make([]Executor, len(v.Inners)),
-		Condition:  b.joinConditions(v.Conditions),
-		fields:     v.Fields(),
-		ctx:        b.ctx,
-		builder:    b,
-	}
-	return e
 }
 
 func (b *executorBuilder) joinConditions(conditions []ast.ExprNode) ast.ExprNode {
@@ -389,117 +185,6 @@ func (b *executorBuilder) buildSelectLock(v *plan.SelectLock) Executor {
 	return e
 }
 
-func (b *executorBuilder) buildSelectFields(v *plan.SelectFields) Executor {
-	src := b.build(v.GetChildByIndex(0))
-	e := &SelectFieldsExec{
-		Src:          src,
-		ResultFields: v.Fields(),
-		ctx:          b.ctx,
-	}
-	return e
-}
-
-func (b *executorBuilder) buildAggregate(v *plan.Aggregate) Executor {
-	src := b.build(v.GetChildByIndex(0))
-	e := &AggregateExec{
-		Src:          src,
-		ResultFields: v.Fields(),
-		ctx:          b.ctx,
-		AggFuncs:     v.AggFuncs,
-		GroupByItems: v.GroupByItems,
-	}
-	xSrc, ok := src.(XExecutor)
-	if !ok {
-		return e
-	}
-	txn, err := b.ctx.GetTxn(false)
-	if err != nil {
-		b.err = err
-		return nil
-	}
-	client := txn.GetClient()
-	if len(v.GroupByItems) > 0 && !client.SupportRequestType(kv.ReqTypeSelect, kv.ReqSubTypeGroupBy) {
-		return e
-	}
-	log.Debugf("Use XAggregateExec with %d aggs", len(v.AggFuncs))
-	// Convert aggregate function exprs to pb.
-	pbAggFuncs := make([]*tipb.Expr, 0, len(v.AggFuncs))
-	for _, af := range v.AggFuncs {
-		if af.Distinct {
-			// We do not support distinct push down.
-			return e
-		}
-		pbAggFunc := b.aggFuncToPBExpr(client, af, xSrc.GetTableName())
-		if pbAggFunc == nil {
-			return e
-		}
-		pbAggFuncs = append(pbAggFuncs, pbAggFunc)
-	}
-	pbByItems := make([]*tipb.ByItem, 0, len(v.GroupByItems))
-	// Convert groupby to pb
-	for _, item := range v.GroupByItems {
-		pbByItem := b.groupByItemToPB(client, item, xSrc.GetTableName())
-		if pbByItem == nil {
-			return e
-		}
-		pbByItems = append(pbByItems, pbByItem)
-	}
-	// compose aggregate info
-	// We should infer fields type.
-	// Each agg item will be splitted into two datums: count and value
-	// The first field should be group key.
-	fields := make([]*types.FieldType, 0, 1+2*len(v.AggFuncs))
-	gk := types.NewFieldType(mysql.TypeBlob)
-	gk.Charset = charset.CharsetBin
-	gk.Collate = charset.CollationBin
-	fields = append(fields, gk)
-	// There will be one or two fields in the result row for each AggregateFuncExpr.
-	// Count needs count partial result field.
-	// Sum, FirstRow, Max, Min, GroupConcat need value partial result field.
-	// Avg needs both count and value partial result field.
-	for _, agg := range v.AggFuncs {
-		name := strings.ToLower(agg.F)
-		if needCount(name) {
-			// count partial result field
-			ft := types.NewFieldType(mysql.TypeLonglong)
-			ft.Flen = 21
-			ft.Charset = charset.CharsetBin
-			ft.Collate = charset.CollationBin
-			fields = append(fields, ft)
-		}
-		if needValue(name) {
-			// value partial result field
-			fields = append(fields, agg.GetType())
-		}
-	}
-	xSrc.AddAggregate(pbAggFuncs, pbByItems, fields)
-	hasGroupBy := len(v.GroupByItems) > 0
-	xe := &XAggregateExec{
-		Src:          src,
-		ResultFields: v.Fields(),
-		ctx:          b.ctx,
-		AggFuncs:     v.AggFuncs,
-		hasGroupBy:   hasGroupBy,
-	}
-	return xe
-}
-
-func (b *executorBuilder) buildHaving(v *plan.Having) Executor {
-	src := b.build(v.GetChildByIndex(0))
-	return b.buildFilter(src, v.Conditions)
-}
-
-func (b *executorBuilder) buildSort(v *plan.Sort) Executor {
-	src := b.build(v.GetChildByIndex(0))
-	e := &SortExec{
-		Src:     src,
-		ByItems: v.ByItems,
-		ctx:     b.ctx,
-		Limit:   v.ExecLimit,
-	}
-	return e
-}
-
 func (b *executorBuilder) buildLimit(v *plan.Limit) Executor {
 	src := b.build(v.GetChildByIndex(0))
 	if x, ok := src.(NewXExecutor); ok {
@@ -512,19 +197,6 @@ func (b *executorBuilder) buildLimit(v *plan.Limit) Executor {
 		Offset: v.Offset,
 		Count:  v.Count,
 		schema: v.GetSchema(),
-	}
-	return e
-}
-
-func (b *executorBuilder) buildUnion(v *plan.Union) Executor {
-	e := &UnionExec{
-		schema: v.GetSchema(),
-		fields: v.Fields(),
-		Sels:   make([]Executor, len(v.Selects)),
-	}
-	for i, sel := range v.Selects {
-		selExec := b.build(sel)
-		e.Sels[i] = selExec
 	}
 	return e
 }
@@ -549,21 +221,6 @@ func (b *executorBuilder) buildExecute(v *plan.Execute) Executor {
 		Name:      v.Name,
 		UsingVars: v.UsingVars,
 		ID:        v.ID,
-	}
-}
-
-func (b *executorBuilder) buildUpdate(v *plan.Update) Executor {
-	selExec := b.build(v.SelectPlan)
-	return &UpdateExec{ctx: b.ctx, SelectExec: selExec, OrderedList: v.OrderedList}
-}
-
-func (b *executorBuilder) buildDelete(v *plan.Delete) Executor {
-	selExec := b.build(v.SelectPlan)
-	return &DeleteExec{
-		ctx:          b.ctx,
-		SelectExec:   selExec,
-		Tables:       v.Tables,
-		IsMultiTable: v.IsMultiTable,
 	}
 }
 
@@ -661,30 +318,6 @@ func (b *executorBuilder) buildExplain(v *plan.Explain) Executor {
 		StmtPlan: v.StmtPlan,
 		fields:   v.Fields(),
 	}
-}
-
-// buildUnionScanExec builds a union scan executor, the src Executor is either
-// *XSelectTableExec or *XSelectIndexExec.
-func (b *executorBuilder) buildUnionScanExec(src Executor) *UnionScanExec {
-	us := &UnionScanExec{ctx: b.ctx, Src: src}
-	switch x := src.(type) {
-	case *XSelectTableExec:
-		us.desc = x.tablePlan.Desc
-		us.dirty = getDirtyDB(b.ctx).getDirtyTable(x.table.Meta().ID)
-		us.condition = b.joinConditions(append(x.tablePlan.AccessConditions, x.tablePlan.FilterConditions...))
-		us.buildAndSortAddedRows(x.table, x.tablePlan.TableAsName)
-	case *XSelectIndexExec:
-		us.desc = x.indexPlan.Desc
-		for _, ic := range x.indexPlan.Index.Columns {
-			us.usedIndex = append(us.usedIndex, ic.Offset)
-		}
-		us.dirty = getDirtyDB(b.ctx).getDirtyTable(x.table.Meta().ID)
-		us.condition = b.joinConditions(append(x.indexPlan.AccessConditions, x.indexPlan.FilterConditions...))
-		us.buildAndSortAddedRows(x.table, x.indexPlan.TableAsName)
-	default:
-		b.err = ErrUnknownPlan.Gen("Unknown Plan %T", src)
-	}
-	return us
 }
 
 func (b *executorBuilder) buildNewUnionScanExec(src Executor, condition expression.Expression) *UnionScanExec {
