+++ executor/executor_xapi.go
@@ -15,17 +15,11 @@ package executor
 
 import (
 	"math"
-	"sort"
 	"strings"
-	"sync"
-	"time"
 
 	"github.com/juju/errors"
-	"github.com/ngaut/log"
 	"github.com/pingcap/tidb/ast"
-	"github.com/pingcap/tidb/context"
 	"github.com/pingcap/tidb/evaluator"
-	"github.com/pingcap/tidb/expression"
 	"github.com/pingcap/tidb/kv"
 	"github.com/pingcap/tidb/model"
 	"github.com/pingcap/tidb/mysql"
@@ -40,114 +34,12 @@ import (
 
 const defaultConcurrency int = 10
 
-var (
-	_ Executor  = &XSelectTableExec{}
-	_ Executor  = &XSelectIndexExec{}
-	_ XExecutor = &XSelectTableExec{}
-	_ XExecutor = &XSelectIndexExec{}
-)
-
 // XExecutor defines some interfaces used by dist-sql.
 type XExecutor interface {
 	// AddAggregate adds aggregate info into an executor.
 	AddAggregate(funcs []*tipb.Expr, byItems []*tipb.ByItem, fields []*types.FieldType)
 	// GetTableName gets the table name of this XExecutor.
 	GetTableName() *ast.TableName
-}
-
-// XSelectTableExec represents XAPI select executor.
-type XSelectTableExec struct {
-	XExecutor
-	table            table.Table
-	tablePlan        *plan.TableScan
-	where            *tipb.Expr
-	ctx              context.Context
-	result           xapi.SelectResult
-	subResult        xapi.PartialResult
-	supportDesc      bool
-	allFiltersPushed bool
-	aggFuncs         []*tipb.Expr
-	byItems          []*tipb.ByItem
-	aggFields        []*types.FieldType
-	// Indicate if the exec is handling aggregate result.
-	aggregate bool
-}
-
-// AddAggregate implements XExecutor interface.
-func (e *XSelectTableExec) AddAggregate(funcs []*tipb.Expr, byItems []*tipb.ByItem, fields []*types.FieldType) {
-	e.aggFuncs = funcs
-	e.byItems = byItems
-	e.aggFields = fields
-	e.aggregate = true
-}
-
-// GetTableName implements XExecutor interface.
-func (e *XSelectTableExec) GetTableName() *ast.TableName {
-	return e.tablePlan.TableName
-}
-
-// Next implements Executor Next interface.
-func (e *XSelectTableExec) Next() (*Row, error) {
-	if e.result == nil {
-		err := e.doRequest()
-		if err != nil {
-			return nil, errors.Trace(err)
-		}
-	}
-	for {
-		if e.subResult == nil {
-			var err error
-			startTs := time.Now()
-			e.subResult, err = e.result.Next()
-			if err != nil {
-				return nil, errors.Trace(err)
-			}
-			if e.subResult == nil {
-				return nil, nil
-			}
-			log.Debugf("[TIME_TABLE_SCAN] %v", time.Now().Sub(startTs))
-		}
-		h, rowData, err := e.subResult.Next()
-		if err != nil {
-			return nil, errors.Trace(err)
-		}
-		if rowData == nil {
-			e.subResult = nil
-			continue
-		}
-		if e.aggregate {
-			// compose aggreagte row
-			return &Row{Data: rowData}, nil
-		}
-		fullRowData := make([]types.Datum, len(e.tablePlan.Fields()))
-		var j int
-		for i, field := range e.tablePlan.Fields() {
-			if field.Referenced {
-				fullRowData[i] = rowData[j]
-				field.Expr.SetDatum(rowData[j])
-				j++
-			}
-		}
-		return resultRowToRow(e.table, h, fullRowData, e.tablePlan.TableAsName), nil
-	}
-}
-
-// Fields implements Executor Fields interface.
-func (e *XSelectTableExec) Fields() []*ast.ResultField {
-	return e.tablePlan.Fields()
-}
-
-// Schema implements Executor Schema interface.
-func (e *XSelectTableExec) Schema() expression.Schema {
-	return nil
-}
-
-// Close implements Executor Close interface.
-func (e *XSelectTableExec) Close() error {
-	if e.result != nil {
-		e.result.Close()
-	}
-	return nil
 }
 
 func resultRowToRow(t table.Table, h int64, data []types.Datum, tableAsName *model.CIStr) *Row {
@@ -157,73 +49,6 @@ func resultRowToRow(t table.Table, h int64, data []types.Datum, tableAsName *mod
 		TableAsName: tableAsName,
 	}
 	return &Row{Data: data, RowKeys: []*RowKeyEntry{entry}}
-}
-
-func (e *XSelectTableExec) doRequest() error {
-	txn, err := e.ctx.GetTxn(false)
-	if err != nil {
-		return errors.Trace(err)
-	}
-	selReq := new(tipb.SelectRequest)
-	selReq.StartTs = txn.StartTS()
-	selReq.Fields = resultFieldsToPBExpression(e.tablePlan.Fields())
-	selReq.Where = e.where
-	selReq.Ranges = tableRangesToPBRanges(e.tablePlan.Ranges)
-	if e.supportDesc {
-		if e.tablePlan.Desc {
-			selReq.OrderBy = append(selReq.OrderBy, &tipb.ByItem{Desc: e.tablePlan.Desc})
-		}
-		// Limit can be pushed only if desc is supported.
-		if e.allFiltersPushed {
-			selReq.Limit = e.tablePlan.LimitCount
-		}
-	}
-	columns := make([]*model.ColumnInfo, 0, len(e.tablePlan.Fields()))
-	for _, v := range e.tablePlan.Fields() {
-		if v.Referenced {
-			columns = append(columns, v.Column)
-		}
-	}
-	selReq.TableInfo = &tipb.TableInfo{
-		TableId: e.table.Meta().ID,
-	}
-	selReq.TableInfo.Columns = xapi.ColumnsToProto(columns, e.table.Meta().PKIsHandle)
-	// Aggregate Info
-	selReq.Aggregates = e.aggFuncs
-	selReq.GroupBy = e.byItems
-	e.result, err = xapi.Select(txn.GetClient(), selReq, defaultConcurrency, true)
-	if err != nil {
-		return errors.Trace(err)
-	}
-	if len(selReq.Aggregates) > 0 || len(selReq.GroupBy) > 0 {
-		// The returned rows should be aggregate partial result.
-		e.result.SetFields(e.aggFields)
-	}
-	return nil
-}
-
-// XSelectIndexExec represents XAPI select index executor.
-type XSelectIndexExec struct {
-	XExecutor
-	table       table.Table
-	indexPlan   *plan.IndexScan
-	ctx         context.Context
-	where       *tipb.Expr
-	supportDesc bool
-
-	byItems []*tipb.ByItem
-	// Aggregate functions.
-	aggFuncs []*tipb.Expr
-	// Aggregate function result fields.
-	aggFields []*types.FieldType
-	// Indicate if the exec is handling aggregate result.
-	aggregate bool
-
-	tasks      []*lookupTableTask
-	taskCursor int
-	indexOrder map[int64]int
-
-	mu sync.Mutex
 }
 
 // BaseLookupTableTaskSize represents base number of handles for a lookupTableTask.
@@ -247,202 +72,6 @@ type lookupTableTask struct {
 	doneCh  chan error
 }
 
-// Schema implements Executor Schema interface.
-func (e *XSelectIndexExec) Schema() expression.Schema {
-	return nil
-}
-
-// AddAggregate implements XExecutor interface.
-func (e *XSelectIndexExec) AddAggregate(funcs []*tipb.Expr, byItems []*tipb.ByItem, fields []*types.FieldType) {
-	e.aggFuncs = funcs
-	e.byItems = byItems
-	e.aggFields = fields
-	e.aggregate = true
-}
-
-// GetTableName implements XExecutor interface.
-func (e *XSelectIndexExec) GetTableName() *ast.TableName {
-	return e.indexPlan.TableName
-}
-
-// Fields implements Executor Fields interface.
-func (e *XSelectIndexExec) Fields() []*ast.ResultField {
-	return e.indexPlan.Fields()
-}
-
-// Next implements Executor Next interface.
-func (e *XSelectIndexExec) Next() (*Row, error) {
-	if e.tasks == nil {
-		startTs := time.Now()
-		handles, err := e.fetchHandles()
-		if err != nil {
-			return nil, errors.Trace(err)
-		}
-		log.Debugf("[TIME_INDEX_SCAN] time: %v handles: %d", time.Now().Sub(startTs), len(handles))
-		e.buildTableTasks(handles)
-		limitCount := int64(-1)
-		if e.indexPlan.LimitCount != nil {
-			limitCount = *e.indexPlan.LimitCount
-		}
-		concurrency := 2
-		if e.indexPlan.NoLimit {
-			concurrency = len(e.tasks)
-		} else if limitCount > int64(BaseLookupTableTaskSize) {
-			concurrency = int(limitCount/int64(BaseLookupTableTaskSize) + 1)
-		}
-		log.Debugf("[TIME_INDEX_TABLE_CONCURRENT_SCAN] start %d workers", concurrency)
-		e.runTableTasks(concurrency)
-	}
-	for {
-		if e.taskCursor >= len(e.tasks) {
-			return nil, nil
-		}
-		task := e.tasks[e.taskCursor]
-		if !task.done {
-			startTs := time.Now()
-			err := <-task.doneCh
-			if err != nil {
-				return nil, errors.Trace(err)
-			}
-			task.done = true
-			log.Debugf("[TIME_INDEX_TABLE_SCAN] time: %v handles: %d", time.Now().Sub(startTs), len(task.handles))
-		}
-		if task.cursor < len(task.rows) {
-			row := task.rows[task.cursor]
-			if !e.aggregate {
-				for i, field := range e.indexPlan.Fields() {
-					field.Expr.SetDatum(row.Data[i])
-				}
-			}
-			task.cursor++
-			return row, nil
-		}
-		e.taskCursor++
-	}
-}
-
-func (e *XSelectIndexExec) pickAndExecTask() {
-	for {
-		// Pick a new task.
-		e.mu.Lock()
-		var task *lookupTableTask
-		for _, t := range e.tasks {
-			if t.status == taskNew {
-				task = t
-				task.status = taskRunning
-				break
-			}
-		}
-		e.mu.Unlock()
-		if task == nil {
-			// No more task to run.
-			break
-		}
-		// Execute the picked task.
-		err := e.executeTask(task)
-		e.mu.Lock()
-		task.status = taskDone
-		e.mu.Unlock()
-		task.doneCh <- err
-	}
-}
-
-func (e *XSelectIndexExec) runTableTasks(n int) {
-	for i := 0; i < n && i < len(e.tasks); i++ {
-		go e.pickAndExecTask()
-	}
-}
-
-func (e *XSelectIndexExec) executeTask(task *lookupTableTask) error {
-	sort.Sort(int64Slice(task.handles))
-	tblResult, err := e.doTableRequest(task.handles)
-	if err != nil {
-		return errors.Trace(err)
-	}
-	task.rows, err = e.extractRowsFromTableResult(e.table, tblResult)
-	if err != nil {
-		return errors.Trace(err)
-	}
-	// TODO: check this
-	if !e.indexPlan.OutOfOrder && !e.aggregate {
-		// Restore the index order.
-		sorter := &rowsSorter{order: e.indexOrder, rows: task.rows}
-		if e.indexPlan.Desc && !e.supportDesc {
-			sort.Sort(sort.Reverse(sorter))
-		} else {
-			sort.Sort(sorter)
-		}
-	}
-	return nil
-}
-
-// Close implements Executor Close interface.
-func (e *XSelectIndexExec) Close() error {
-	return nil
-}
-
-// fetchHandle fetches all handles from the index.
-// This should be optimized to fetch handles in batch.
-func (e *XSelectIndexExec) fetchHandles() ([]int64, error) {
-	idxResult, err := e.doIndexRequest()
-	if err != nil {
-		return nil, errors.Trace(err)
-	}
-	handles, err := extractHandlesFromIndexResult(idxResult)
-	if err != nil {
-		return nil, errors.Trace(err)
-	}
-	// TODO: check this
-	if !e.indexPlan.OutOfOrder && !e.aggregate {
-		// Save the index order.
-		e.indexOrder = make(map[int64]int)
-		for i, h := range handles {
-			e.indexOrder[h] = i
-		}
-	}
-	return handles, nil
-}
-
-func (e *XSelectIndexExec) buildTableTasks(handles []int64) {
-	// Build tasks with increasing batch size.
-	var taskSizes []int
-	total := len(handles)
-	batchSize := BaseLookupTableTaskSize
-	for total > 0 {
-		if batchSize > total {
-			batchSize = total
-		}
-		taskSizes = append(taskSizes, batchSize)
-		total -= batchSize
-		if batchSize < MaxLookupTableTaskSize {
-			batchSize *= 2
-		}
-	}
-	if e.indexPlan.Desc && !e.supportDesc {
-		// Reverse tasks sizes.
-		for i := 0; i < len(taskSizes)/2; i++ {
-			j := len(taskSizes) - i - 1
-			taskSizes[i], taskSizes[j] = taskSizes[j], taskSizes[i]
-		}
-	}
-	e.tasks = make([]*lookupTableTask, len(taskSizes))
-	for i, size := range taskSizes {
-		task := &lookupTableTask{
-			handles: handles[:size],
-		}
-		task.doneCh = make(chan error, 1)
-		handles = handles[size:]
-		e.tasks[i] = task
-	}
-	if e.indexPlan.Desc && !e.supportDesc {
-		// Reverse tasks order.
-		for i := 0; i < len(e.tasks)/2; i++ {
-			j := len(e.tasks) - i - 1
-			e.tasks[i], e.tasks[j] = e.tasks[j], e.tasks[i]
-		}
-	}
-}
-
 type rowsSorter struct {
 	order map[int64]int
 	rows  []*Row
@@ -460,104 +89,6 @@ func (s *rowsSorter) Len() int {
 
 func (s *rowsSorter) Swap(i, j int) {
 	s.rows[i], s.rows[j] = s.rows[j], s.rows[i]
-}
-
-func (e *XSelectIndexExec) doIndexRequest() (xapi.SelectResult, error) {
-	txn, err := e.ctx.GetTxn(false)
-	if err != nil {
-		return nil, errors.Trace(err)
-	}
-	selIdxReq := new(tipb.SelectRequest)
-	selIdxReq.StartTs = txn.StartTS()
-	selIdxReq.IndexInfo = xapi.IndexToProto(e.table.Meta(), e.indexPlan.Index)
-	if len(e.indexPlan.FilterConditions) == 0 {
-		// Push limit to index request only if there is not filter conditions.
-		selIdxReq.Limit = e.indexPlan.LimitCount
-	}
-	if e.indexPlan.Desc {
-		selIdxReq.OrderBy = append(selIdxReq.OrderBy, &tipb.ByItem{Desc: e.indexPlan.Desc})
-	}
-	fieldTypes := make([]*types.FieldType, len(e.indexPlan.Index.Columns))
-	for i, v := range e.indexPlan.Index.Columns {
-		fieldTypes[i] = &(e.table.Cols()[v.Offset].FieldType)
-	}
-	selIdxReq.Ranges, err = indexRangesToPBRanges(e.indexPlan.Ranges, fieldTypes)
-	if err != nil {
-		return nil, errors.Trace(err)
-	}
-	concurrency := 1
-	if e.indexPlan.OutOfOrder {
-		concurrency = 10
-	}
-	return xapi.Select(txn.GetClient(), selIdxReq, concurrency, !e.indexPlan.OutOfOrder)
-}
-
-func (e *XSelectIndexExec) doTableRequest(handles []int64) (xapi.SelectResult, error) {
-	txn, err := e.ctx.GetTxn(false)
-	if err != nil {
-		return nil, errors.Trace(err)
-	}
-	// The handles are not in original index order, so we can't push limit here.
-	selTableReq := new(tipb.SelectRequest)
-	selTableReq.StartTs = txn.StartTS()
-	columns := make([]*model.ColumnInfo, 0, len(e.indexPlan.Fields()))
-	for _, v := range e.indexPlan.Fields() {
-		if v.Referenced {
-			columns = append(columns, v.Column)
-		}
-	}
-	selTableReq.TableInfo = &tipb.TableInfo{
-		TableId: e.table.Meta().ID,
-	}
-	selTableReq.TableInfo.Columns = xapi.ColumnsToProto(columns, e.table.Meta().PKIsHandle)
-	selTableReq.Fields = resultFieldsToPBExpression(e.indexPlan.Fields())
-	for _, h := range handles {
-		if h == math.MaxInt64 {
-			// We can't convert MaxInt64 into an left closed, right open range.
-			continue
-		}
-		pbRange := new(tipb.KeyRange)
-		pbRange.Low = codec.EncodeInt(nil, h)
-		pbRange.High = kv.Key(pbRange.Low).PrefixNext()
-		selTableReq.Ranges = append(selTableReq.Ranges, pbRange)
-	}
-	selTableReq.Where = e.where
-	// Aggregate Info
-	selTableReq.Aggregates = e.aggFuncs
-	selTableReq.GroupBy = e.byItems
-	resp, err := xapi.Select(txn.GetClient(), selTableReq, defaultConcurrency, true)
-	if err != nil {
-		return nil, errors.Trace(err)
-	}
-	if e.aggregate {
-		// The returned rows should be aggregate partial result.
-		resp.SetFields(e.aggFields)
-	}
-	return resp, nil
-}
-
-// conditionsToPBExpr tries to convert filter conditions to a tipb.Expr.
-// not supported conditions will be returned in remained.
-func (b *executorBuilder) conditionsToPBExpr(client kv.Client, exprs []ast.ExprNode, tn *ast.TableName) (pbexpr *tipb.Expr,
-	remained []ast.ExprNode) {
-	for _, expr := range exprs {
-		v := b.exprToPBExpr(client, expr, tn)
-		if v == nil {
-			remained = append(remained, expr)
-		} else {
-			if pbexpr == nil {
-				pbexpr = v
-			} else {
-				// merge multiple converted pb expression into an AND expression.
-				pbexpr = &tipb.Expr{Tp: tipb.ExprType_And, Children: []*tipb.Expr{pbexpr, v}}
-			}
-		}
-	}
-	return
-}
-
-func resultFieldsToPBExpression(fields []*ast.ResultField) []*tipb.Expr {
-	return nil
 }
 
 func tableRangesToPBRanges(tableRanges []plan.TableRange) []*tipb.KeyRange {
@@ -694,55 +225,6 @@ func extractHandlesFromIndexSubResult(subResult xapi.PartialResult) ([]int64, er
 		handles = append(handles, h)
 	}
 	return handles, nil
-}
-
-func (e *XSelectIndexExec) extractRowsFromTableResult(t table.Table, tblResult xapi.SelectResult) ([]*Row, error) {
-	var rows []*Row
-	for {
-		subResult, err := tblResult.Next()
-		if err != nil {
-			return nil, errors.Trace(err)
-		}
-		if subResult == nil {
-			break
-		}
-		subRows, err := e.extractRowsFromSubResult(t, subResult)
-		if err != nil {
-			return nil, errors.Trace(err)
-		}
-		rows = append(rows, subRows...)
-	}
-	return rows, nil
-}
-
-func (e *XSelectIndexExec) extractRowsFromSubResult(t table.Table, subResult xapi.PartialResult) ([]*Row, error) {
-	var rows []*Row
-	for {
-		h, rowData, err := subResult.Next()
-		if err != nil {
-			return nil, errors.Trace(err)
-		}
-		if rowData == nil {
-			break
-		}
-		if e.aggregate {
-			// compose aggreagte row
-			row := &Row{Data: rowData}
-			rows = append(rows, row)
-			continue
-		}
-		fullRowData := make([]types.Datum, len(e.indexPlan.Fields()))
-		var j int
-		for i, field := range e.indexPlan.Fields() {
-			if field.Referenced {
-				fullRowData[i] = rowData[j]
-				j++
-			}
-		}
-		row := resultRowToRow(t, h, fullRowData, e.indexPlan.TableAsName)
-		rows = append(rows, row)
-	}
-	return rows, nil
 }
 
 type int64Slice []int64
