+++ plan/planbuilder.go
@@ -14,11 +14,7 @@
 package plan
 
 import (
-	"math"
-
-	"github.com/golang/protobuf/proto"
 	"github.com/juju/errors"
-	"github.com/ngaut/log"
 	"github.com/pingcap/tidb/ast"
 	"github.com/pingcap/tidb/context"
 	"github.com/pingcap/tidb/expression"
@@ -45,8 +41,8 @@ const (
 
 // BuildPlan builds a plan from a node.
 // It returns ErrUnsupportedType if ast.Node type is not supported yet.
-func BuildPlan(node ast.Node, sb SubQueryBuilder) (Plan, error) {
-	builder := planBuilder{sb: sb, allocator: new(idAllocator)}
+func BuildPlan(node ast.Node) (Plan, error) {
+	builder := planBuilder{allocator: new(idAllocator)}
 	p := builder.build(node)
 	return p, builder.err
 }
@@ -56,7 +52,6 @@ func BuildPlan(node ast.Node, sb SubQueryBuilder) (Plan, error) {
 type planBuilder struct {
 	err          error
 	hasAgg       bool
-	sb           SubQueryBuilder
 	obj          interface{}
 	allocator    *idAllocator
 	ctx          context.Context
@@ -85,10 +80,7 @@ func (b *planBuilder) build(node ast.Node) Plan {
 	case *ast.DeallocateStmt:
 		return &Deallocate{Name: x.Name}
 	case *ast.DeleteStmt:
-		if UseNewPlanner {
-			return b.buildNewDelete(x)
-		}
-		return b.buildDelete(x)
+		return b.buildNewDelete(x)
 	case *ast.DropDatabaseStmt:
 		return b.buildDDL(x)
 	case *ast.DropIndexStmt:
@@ -104,20 +96,11 @@ func (b *planBuilder) build(node ast.Node) Plan {
 	case *ast.PrepareStmt:
 		return b.buildPrepare(x)
 	case *ast.SelectStmt:
-		if UseNewPlanner {
-			return b.buildNewSelect(x)
-		}
-		return b.buildSelect(x)
+		return b.buildNewSelect(x)
 	case *ast.UnionStmt:
-		if UseNewPlanner {
-			return b.buildNewUnion(x)
-		}
-		return b.buildUnion(x)
+		return b.buildNewUnion(x)
 	case *ast.UpdateStmt:
-		if UseNewPlanner {
-			return b.buildNewUpdate(x)
-		}
-		return b.buildUpdate(x)
+		return b.buildNewUpdate(x)
 	case *ast.UseStmt:
 		return b.buildSimple(x)
 	case *ast.SetStmt:
@@ -220,153 +203,6 @@ func (b *planBuilder) extractSelectAgg(sel *ast.SelectStmt) []*ast.AggregateFunc
 	return extractor.AggFuncs
 }
 
-func (b *planBuilder) buildSubquery(n ast.Node) {
-	sv := &subqueryVisitor{
-		builder: b,
-	}
-	_, ok := n.Accept(sv)
-	if !ok {
-		log.Errorf("Extract subquery error")
-	}
-}
-
-func (b *planBuilder) buildSelect(sel *ast.SelectStmt) Plan {
-	var aggFuncs []*ast.AggregateFuncExpr
-	hasAgg := b.detectSelectAgg(sel)
-	canPushLimit := !hasAgg
-	if hasAgg {
-		aggFuncs = b.extractSelectAgg(sel)
-	}
-	// Build subquery
-	// Convert subquery to expr with plan
-	b.buildSubquery(sel)
-	var p Plan
-	if sel.From != nil {
-		p = b.buildFrom(sel)
-		if b.err != nil {
-			return nil
-		}
-		if sel.LockTp != ast.SelectLockNone {
-			p = b.buildSelectLock(p, sel.LockTp)
-			if b.err != nil {
-				return nil
-			}
-		}
-		if hasAgg {
-			p = b.buildAggregate(p, aggFuncs, sel.GroupBy)
-		}
-		p = b.buildSelectFields(p, sel.GetResultFields())
-		if b.err != nil {
-			return nil
-		}
-	} else {
-		canPushLimit = false
-		if sel.Where != nil {
-			p = b.buildTableDual(sel)
-		}
-		if hasAgg {
-			p = b.buildAggregate(p, aggFuncs, nil)
-		}
-		p = b.buildSelectFields(p, sel.GetResultFields())
-		if b.err != nil {
-			return nil
-		}
-	}
-	if sel.Having != nil {
-		p = b.buildHaving(p, sel.Having)
-		if b.err != nil {
-			return nil
-		}
-	}
-	if sel.Distinct {
-		canPushLimit = false
-		p = b.buildDistinct(p)
-		if b.err != nil {
-			return nil
-		}
-	}
-	if sel.OrderBy != nil && !pushOrder(p, sel.OrderBy.Items) {
-		canPushLimit = false
-		p = b.buildSort(p, sel.OrderBy.Items)
-		if b.err != nil {
-			return nil
-		}
-	}
-	if sel.Limit != nil {
-		if canPushLimit {
-			pushLimit(p, sel.Limit)
-		}
-		p = b.buildLimit(p, sel.Limit)
-		if b.err != nil {
-			return nil
-		}
-	}
-	return p
-}
-
-func (b *planBuilder) buildFrom(sel *ast.SelectStmt) Plan {
-	from := sel.From.TableRefs
-	if from.Right == nil {
-		return b.buildTableSource(sel)
-	}
-	return b.buildJoin(sel)
-}
-
-func (b *planBuilder) buildTableSource(sel *ast.SelectStmt) Plan {
-	from := sel.From.TableRefs
-	ts, ok := from.Left.(*ast.TableSource)
-	if !ok {
-		b.err = ErrUnsupportedType.Gen("Unsupported type %T", from.Left)
-		return nil
-	}
-	var bestPlan Plan
-	switch v := ts.Source.(type) {
-	case *ast.TableName:
-	case *ast.SelectStmt:
-		bestPlan = b.buildSelect(v)
-	case *ast.UnionStmt:
-		bestPlan = b.buildUnion(v)
-	}
-	if bestPlan != nil {
-		return bestPlan
-	}
-	tn, ok := ts.Source.(*ast.TableName)
-	if !ok {
-		b.err = ErrUnsupportedType.Gen("Unsupported type %T", ts.Source)
-		return nil
-	}
-	conditions := splitWhere(sel.Where)
-	path := &joinPath{table: tn, conditions: conditions}
-	candidates := b.buildAllAccessMethodsPlan(path)
-	var lowestCost float64
-	for _, v := range candidates {
-		cost := EstimateCost(b.buildPseudoSelectPlan(v, sel))
-		if bestPlan == nil {
-			bestPlan = v
-			lowestCost = cost
-		}
-		if cost <= lowestCost {
-			bestPlan = v
-			lowestCost = cost
-		}
-	}
-	return bestPlan
-}
-
-func (b *planBuilder) buildAllAccessMethodsPlan(path *joinPath) []Plan {
-	indices, includeTableScan := availableIndices(path.table)
-	var candidates []Plan
-	if includeTableScan {
-		p := b.buildTableScanPlan(path)
-		candidates = append(candidates, p)
-	}
-	for _, index := range indices {
-		ip := b.buildIndexScanPlan(index, path)
-		candidates = append(candidates, ip)
-	}
-	return candidates
-}
-
 func availableIndices(table *ast.TableName) (indices []*model.IndexInfo, includeTableScan bool) {
 	var usableHints []*ast.IndexHint
 	for _, hint := range table.IndexHints {
@@ -443,141 +279,6 @@ func findIndexByName(indices []*model.IndexInfo, name model.CIStr) *model.IndexI
 	return nil
 }
 
-func (b *planBuilder) buildTableDual(sel *ast.SelectStmt) Plan {
-	dual := &TableDual{FilterConditions: splitWhere(sel.Where)}
-	ret := ast.ResultField{}
-	dual.SetFields([]*ast.ResultField{&ret})
-	return dual
-}
-
-func getTableAsName(fields []*ast.ResultField) *model.CIStr {
-	if len(fields) > 0 {
-		return &fields[0].TableAsName
-	}
-	return nil
-}
-
-func (b *planBuilder) buildTableScanPlan(path *joinPath) Plan {
-	tn := path.table
-	p := &TableScan{
-		Table:     tn.TableInfo,
-		TableName: tn,
-	}
-	// Equal condition contains a column from previous joined table.
-	p.RefAccess = len(path.eqConds) > 0
-	p.SetFields(tn.GetResultFields())
-	p.TableAsName = getTableAsName(p.Fields())
-	var pkName model.CIStr
-	if p.Table.PKIsHandle {
-		for _, colInfo := range p.Table.Columns {
-			if mysql.HasPriKeyFlag(colInfo.Flag) {
-				pkName = colInfo.Name
-			}
-		}
-	}
-	for _, con := range path.conditions {
-		if pkName.L != "" {
-			checker := conditionChecker{tableName: tn.TableInfo.Name, pkName: pkName}
-			if checker.check(con) {
-				p.AccessConditions = append(p.AccessConditions, con)
-			} else {
-				p.FilterConditions = append(p.FilterConditions, con)
-			}
-		} else {
-			p.FilterConditions = append(p.FilterConditions, con)
-		}
-	}
-	return p
-}
-
-func (b *planBuilder) buildIndexScanPlan(index *model.IndexInfo, path *joinPath) Plan {
-	tn := path.table
-	ip := &IndexScan{Table: tn.TableInfo, Index: index, TableName: tn}
-	ip.RefAccess = len(path.eqConds) > 0
-	ip.SetFields(tn.GetResultFields())
-	ip.TableAsName = getTableAsName(ip.Fields())
-
-	condMap := map[ast.ExprNode]bool{}
-	for _, con := range path.conditions {
-		condMap[con] = true
-	}
-out:
-	// Build equal access conditions first.
-	// Starts from the first index column, if equal condition is found, add it to access conditions,
-	// proceed to the next index column. until we can't find any equal condition for the column.
-	for ip.AccessEqualCount < len(index.Columns) {
-		for con := range condMap {
-			binop, ok := con.(*ast.BinaryOperationExpr)
-			if !ok || binop.Op != opcode.EQ {
-				continue
-			}
-			if ast.IsPreEvaluable(binop.L) {
-				binop.L, binop.R = binop.R, binop.L
-			}
-			if !ast.IsPreEvaluable(binop.R) {
-				continue
-			}
-			cn, ok2 := binop.L.(*ast.ColumnNameExpr)
-			if !ok2 || cn.Refer.Column.Name.L != index.Columns[ip.AccessEqualCount].Name.L {
-				continue
-			}
-			ip.AccessConditions = append(ip.AccessConditions, con)
-			delete(condMap, con)
-			ip.AccessEqualCount++
-			continue out
-		}
-		break
-	}
-
-	for con := range condMap {
-		if ip.AccessEqualCount < len(ip.Index.Columns) {
-			// Try to add non-equal access condition for index column at AccessEqualCount.
-			checker := conditionChecker{tableName: tn.TableInfo.Name, idx: index, columnOffset: ip.AccessEqualCount}
-			if checker.check(con) {
-				ip.AccessConditions = append(ip.AccessConditions, con)
-			} else {
-				ip.FilterConditions = append(ip.FilterConditions, con)
-			}
-		} else {
-			ip.FilterConditions = append(ip.FilterConditions, con)
-		}
-	}
-	return ip
-}
-
-// buildPseudoSelectPlan pre-builds more complete plans that may affect total cost.
-// Also set OutOfOrder and NoLimit property.
-func (b *planBuilder) buildPseudoSelectPlan(p Plan, sel *ast.SelectStmt) Plan {
-	if sel.OrderBy == nil {
-		return p
-	}
-	if sel.GroupBy != nil {
-		return p
-	}
-	if !pushOrder(p, sel.OrderBy.Items) {
-		switch x := p.(type) {
-		case *IndexScan:
-			x.OutOfOrder = true
-			x.NoLimit = true
-		}
-		np := &Sort{ByItems: sel.OrderBy.Items}
-		addChild(np, p)
-		p = np
-	}
-	if sel.Limit != nil {
-		np := &Limit{Offset: sel.Limit.Offset, Count: sel.Limit.Count}
-		addChild(np, p)
-		np.SetLimit(0)
-		p = np
-	} else {
-		switch x := p.(type) {
-		case *IndexScan:
-			x.NoLimit = true
-		}
-	}
-	return p
-}
-
 func (b *planBuilder) buildSelectLock(src Plan, lock ast.SelectLockType) *SelectLock {
 	selectLock := &SelectLock{
 		Lock:            lock,
@@ -588,61 +289,6 @@ func (b *planBuilder) buildSelectLock(src Plan, lock ast.SelectLockType) *Select
 	selectLock.SetSchema(src.GetSchema())
 	selectLock.SetFields(src.Fields())
 	return selectLock
-}
-
-func (b *planBuilder) buildSelectFields(src Plan, fields []*ast.ResultField) Plan {
-	selectFields := &SelectFields{}
-	addChild(selectFields, src)
-	selectFields.SetFields(fields)
-	return selectFields
-}
-
-func (b *planBuilder) buildAggregate(src Plan, aggFuncs []*ast.AggregateFuncExpr, groupby *ast.GroupByClause) Plan {
-	// Add aggregate plan.
-	aggPlan := &Aggregate{
-		AggFuncs: aggFuncs,
-	}
-	addChild(aggPlan, src)
-	if src != nil {
-		aggPlan.SetFields(src.Fields())
-	}
-	if groupby != nil {
-		aggPlan.GroupByItems = groupby.Items
-	}
-	return aggPlan
-}
-
-func (b *planBuilder) buildHaving(src Plan, having *ast.HavingClause) Plan {
-	p := &Having{
-		Conditions: splitWhere(having.Expr),
-	}
-	addChild(p, src)
-	p.SetFields(src.Fields())
-	return p
-}
-
-func (b *planBuilder) buildSort(src Plan, byItems []*ast.ByItem) Plan {
-	sort := &Sort{
-		ByItems: byItems,
-	}
-	addChild(sort, src)
-	sort.SetFields(src.Fields())
-	return sort
-}
-
-func (b *planBuilder) buildLimit(src Plan, limit *ast.Limit) Plan {
-	li := &Limit{
-		Offset: limit.Offset,
-		Count:  limit.Count,
-	}
-	if s, ok := src.(*Sort); ok {
-		s.ExecLimit = li
-		return s
-	}
-	addChild(li, src)
-	li.SetFields(src.Fields())
-	li.SetSchema(src.GetSchema().DeepCopy())
-	return li
 }
 
 func (b *planBuilder) buildPrepare(x *ast.PrepareStmt) Plan {
@@ -718,99 +364,6 @@ func buildResultField(tableName, name string, tp byte, size int) *ast.ResultFiel
 	}
 }
 
-func pushLimit(p Plan, limit *ast.Limit) {
-	switch x := p.(type) {
-	case *IndexScan:
-		limitCount := limit.Offset + limit.Count
-		if limitCount < math.MaxInt64 {
-			x.LimitCount = proto.Int64(int64(limitCount))
-		}
-	case *TableScan:
-		limitCount := limit.Offset + limit.Count
-		if limitCount < math.MaxInt64 {
-			x.LimitCount = proto.Int64(int64(limitCount))
-		}
-	default:
-		child := x.GetChildByIndex(0)
-		if child != nil {
-			pushLimit(child, limit)
-		}
-	}
-}
-
-// pushOrder tries to push order by items to the plan, returns true if
-// order is pushed.
-func pushOrder(p Plan, items []*ast.ByItem) bool {
-	switch x := p.(type) {
-	case *Aggregate:
-		return false
-	case *IndexScan:
-		if len(items) > len(x.Index.Columns) {
-			return false
-		}
-		var hasDesc bool
-		var hasAsc bool
-		for i, item := range items {
-			var rf *ast.ResultField
-			switch y := item.Expr.(type) {
-			case *ast.ColumnNameExpr:
-				rf = y.Refer
-			case *ast.PositionExpr:
-				rf = y.Refer
-			default:
-				return false
-			}
-			if rf.Table.Name.L != x.Table.Name.L || rf.Column.Name.L != x.Index.Columns[i].Name.L {
-				return false
-			}
-			if item.Desc {
-				if hasAsc {
-					return false
-				}
-				hasDesc = true
-			} else {
-				if hasDesc {
-					return false
-				}
-				hasAsc = true
-			}
-		}
-		x.Desc = hasDesc
-		return true
-	case *TableScan:
-		if len(items) != 1 || !x.Table.PKIsHandle {
-			return false
-		}
-		var refer *ast.ResultField
-		switch x := items[0].Expr.(type) {
-		case *ast.ColumnNameExpr:
-			refer = x.Refer
-		case *ast.PositionExpr:
-			refer = x.Refer
-		default:
-			return false
-		}
-		if mysql.HasPriKeyFlag(refer.Column.Flag) {
-			x.Desc = items[0].Desc
-			return true
-		}
-		return false
-	case *JoinOuter:
-		return false
-	case *JoinInner:
-		return false
-	case *Sort:
-		// Sort plan should not be checked here as there should only be one sort plan in a plan tree.
-		return false
-	default:
-		child := x.GetChildByIndex(0)
-		if child != nil {
-			return pushOrder(child, items)
-		}
-	}
-	return false
-}
-
 // splitWhere split a where expression to a list of AND conditions.
 func splitWhere(where ast.ExprNode) []ast.ExprNode {
 	var conditions []ast.ExprNode
@@ -831,122 +384,11 @@ func splitWhere(where ast.ExprNode) []ast.ExprNode {
 	return conditions
 }
 
-// SubQueryBuilder is the interface for building SubQuery executor.
-type SubQueryBuilder interface {
-	Build(p Plan) ast.SubqueryExec
-}
-
-// subqueryVisitor visits AST and handles SubqueryExpr.
-type subqueryVisitor struct {
-	builder *planBuilder
-}
-
-func (se *subqueryVisitor) Enter(in ast.Node) (out ast.Node, skipChildren bool) {
-	switch x := in.(type) {
-	case *ast.SubqueryExpr:
-		p := se.builder.build(x.Query)
-		// The expr pointer is copied into ResultField when running name resolver.
-		// So we can not just replace the expr node in AST. We need to put SubQuery into the expr.
-		// See optimizer.nameResolver.createResultFields()
-		x.SubqueryExec = se.builder.sb.Build(p)
-		return in, true
-	case *ast.Join:
-		// SubSelect in from clause will be handled in buildJoin().
-		return in, true
-	}
-	return in, false
-}
-
-func (se *subqueryVisitor) Leave(in ast.Node) (out ast.Node, ok bool) {
-	return in, true
-}
-
-func (b *planBuilder) buildUnion(union *ast.UnionStmt) Plan {
-	sels := make([]Plan, len(union.SelectList.Selects))
-	for i, sel := range union.SelectList.Selects {
-		sels[i] = b.buildSelect(sel)
-	}
-	var p Plan
-	p = &Union{
-		Selects: sels,
-	}
-	unionFields := union.GetResultFields()
-	for _, sel := range sels {
-		for i, f := range sel.Fields() {
-			if i == len(unionFields) {
-				b.err = errors.New("The used SELECT statements have a different number of columns")
-				return nil
-			}
-			uField := unionFields[i]
-			/*
-			 * The lengths of the columns in the UNION result take into account the values retrieved by all of the SELECT statements
-			 * SELECT REPEAT('a',1) UNION SELECT REPEAT('b',10);
-			 * +---------------+
-			 * | REPEAT('a',1) |
-			 * +---------------+
-			 * | a             |
-			 * | bbbbbbbbbb    |
-			 * +---------------+
-			 */
-			if f.Column.Flen > uField.Column.Flen {
-				uField.Column.Flen = f.Column.Flen
-			}
-			// For select nul union select "abc", we should not convert "abc" to nil.
-			// And the result field type should be VARCHAR.
-			if uField.Column.Tp == 0 || uField.Column.Tp == mysql.TypeNull {
-				uField.Column.Tp = f.Column.Tp
-			}
-		}
-		addChild(p, sel)
-	}
-	for _, v := range unionFields {
-		v.Expr.SetType(&v.Column.FieldType)
-	}
-
-	p.SetFields(unionFields)
-	if union.Distinct {
-		p = b.buildDistinct(p)
-	}
-	if union.OrderBy != nil {
-		p = b.buildSort(p, union.OrderBy.Items)
-	}
-	if union.Limit != nil {
-		p = b.buildLimit(p, union.Limit)
-	}
-	return p
-}
-
 func (b *planBuilder) buildDistinct(src Plan) Plan {
 	d := &Distinct{}
 	addChild(d, src)
 	d.SetFields(src.Fields())
 	return d
-}
-
-func (b *planBuilder) buildUpdate(update *ast.UpdateStmt) Plan {
-	sel := &ast.SelectStmt{From: update.TableRefs, Where: update.Where, OrderBy: update.Order, Limit: update.Limit}
-	p := b.buildFrom(sel)
-	for _, v := range p.Fields() {
-		v.Referenced = true
-	}
-	if sel.OrderBy != nil && !pushOrder(p, sel.OrderBy.Items) {
-		p = b.buildSort(p, sel.OrderBy.Items)
-		if b.err != nil {
-			return nil
-		}
-	}
-	if sel.Limit != nil {
-		pushLimit(p, sel.Limit)
-		p = b.buildLimit(p, sel.Limit)
-		if b.err != nil {
-			return nil
-		}
-	}
-	orderedList := b.buildUpdateLists(update.List, p.Fields())
-	if b.err != nil {
-		return nil
-	}
-	return &Update{OrderedList: orderedList, SelectPlan: p, basePlan: basePlan{children: []Plan{p}}}
 }
 
 func (b *planBuilder) buildUpdateLists(list []*ast.Assignment, fields []*ast.ResultField) []*ast.Assignment {
@@ -960,37 +402,6 @@ func (b *planBuilder) buildUpdateLists(list []*ast.Assignment, fields []*ast.Res
 		newList[offset] = assign
 	}
 	return newList
-}
-
-func (b *planBuilder) buildDelete(del *ast.DeleteStmt) Plan {
-	sel := &ast.SelectStmt{From: del.TableRefs, Where: del.Where, OrderBy: del.Order, Limit: del.Limit}
-	p := b.buildFrom(sel)
-	for _, v := range p.Fields() {
-		v.Referenced = true
-	}
-	if sel.OrderBy != nil && !pushOrder(p, sel.OrderBy.Items) {
-		p = b.buildSort(p, sel.OrderBy.Items)
-		if b.err != nil {
-			return nil
-		}
-	}
-	if sel.Limit != nil {
-		pushLimit(p, sel.Limit)
-		p = b.buildLimit(p, sel.Limit)
-		if b.err != nil {
-			return nil
-		}
-	}
-	var tables []*ast.TableName
-	if del.Tables != nil {
-		tables = del.Tables.Tables
-	}
-	return &Delete{
-		Tables:       tables,
-		IsMultiTable: del.IsMultiTable,
-		SelectPlan:   p,
-		basePlan:     basePlan{children: []Plan{p}},
-	}
 }
 
 func columnOffsetInFields(cn *ast.ColumnName, fields []*ast.ResultField) (int, error) {
