+++ plan/plans.go
@@ -16,7 +16,6 @@ package plan
 import (
 	"fmt"
 	"github.com/pingcap/tidb/ast"
-	"github.com/pingcap/tidb/model"
 	"github.com/pingcap/tidb/util/types"
 	"strings"
 )
@@ -25,41 +24,6 @@ import (
 type TableRange struct {
 	LowVal  int64
 	HighVal int64
-}
-
-// TableDual represents a dual table plan.
-type TableDual struct {
-	basePlan
-
-	HasAgg bool
-	// FilterConditions can be used to filter result.
-	FilterConditions []ast.ExprNode
-}
-
-// TableScan represents a table scan plan.
-type TableScan struct {
-	basePlan
-
-	Table  *model.TableInfo
-	Desc   bool
-	Ranges []TableRange
-
-	// RefAccess indicates it references a previous joined table, used in explain.
-	RefAccess bool
-
-	// AccessConditions can be used to build index range.
-	AccessConditions []ast.ExprNode
-
-	// FilterConditions can be used to filter result.
-	FilterConditions []ast.ExprNode
-
-	// TableName is used to distinguish the same table selected multiple times in different place,
-	// like 'select * from t where exists(select 1 from t as x where t.c < x.c)'
-	TableName *ast.TableName
-
-	TableAsName *model.CIStr
-
-	LimitCount *int64
 }
 
 // ShowDDL is for showing DDL information.
@@ -131,112 +95,11 @@ func (ir *IndexRange) String() string {
 	return l + strings.Join(lowStrs, " ") + "," + strings.Join(highStrs, " ") + r
 }
 
-// IndexScan represents an index scan plan.
-type IndexScan struct {
-	basePlan
-
-	// The index used.
-	Index *model.IndexInfo
-
-	// The table to lookup.
-	Table *model.TableInfo
-
-	// Ordered and non-overlapping ranges to be scanned.
-	Ranges []*IndexRange
-
-	// Desc indicates whether the index should be scanned in descending order.
-	Desc bool
-
-	// RefAccess indicates it references a previous joined table, used in explain.
-	RefAccess bool
-
-	// AccessConditions can be used to build index range.
-	AccessConditions []ast.ExprNode
-
-	// Number of leading equal access condition.
-	// The offset of each equal condition correspond to the offset of index column.
-	// For example, an index has column (a, b, c), condition is 'a = 0 and b = 0 and c > 0'
-	// AccessEqualCount would be 2.
-	AccessEqualCount int
-
-	// FilterConditions can be used to filter result.
-	FilterConditions []ast.ExprNode
-
-	// OutOfOrder indicates if the index scan can return out of order.
-	OutOfOrder bool
-
-	// NoLimit indicates that this plan need fetch all the rows.
-	NoLimit bool
-
-	// TableName is used to distinguish the same table selected multiple times in different place,
-	// like 'select * from t where exists(select 1 from t as x where t.c < x.c)'
-	TableName *ast.TableName
-
-	TableAsName *model.CIStr
-
-	LimitCount *int64
-}
-
-// JoinOuter represents outer join plan.
-type JoinOuter struct {
-	basePlan
-
-	Outer Plan
-	Inner Plan
-}
-
-// JoinInner represents inner join plan.
-type JoinInner struct {
-	basePlan
-
-	Inners     []Plan
-	Conditions []ast.ExprNode
-}
-
-func (p *JoinInner) String() string {
-	return fmt.Sprintf("JoinInner()")
-}
-
 // SelectLock represents a select lock plan.
 type SelectLock struct {
 	baseLogicalPlan
 
 	Lock ast.SelectLockType
-}
-
-// SetLimit implements Plan SetLimit interface.
-func (p *SelectLock) SetLimit(limit float64) {
-	p.limit = limit
-	p.GetChildByIndex(0).SetLimit(p.limit)
-}
-
-// SelectFields represents a select fields plan.
-type SelectFields struct {
-	basePlan
-}
-
-// SetLimit implements Plan SetLimit interface.
-func (p *SelectFields) SetLimit(limit float64) {
-	p.limit = limit
-	if p.GetChildByIndex(0) != nil {
-		p.GetChildByIndex(0).SetLimit(limit)
-	}
-}
-
-// Sort represents a sorting plan.
-type Sort struct {
-	basePlan
-
-	ByItems []*ast.ByItem
-
-	ExecLimit *Limit
-}
-
-// SetLimit implements Plan SetLimit interface.
-// It set the Src limit only if it is bypassed.
-// Bypass has to be determined before this get called.
-func (p *Sort) SetLimit(limit float64) {
-	p.limit = limit
 }
 
 // Limit represents offset and limit plan.
@@ -247,32 +110,9 @@ type Limit struct {
 	Count  uint64
 }
 
-// SetLimit implements Plan SetLimit interface.
-// As Limit itself determine the real limit,
-// We just ignore the input, and set the real limit.
-func (p *Limit) SetLimit(limit float64) {
-	p.limit = float64(p.Offset + p.Count)
-	p.GetChildByIndex(0).SetLimit(p.limit)
-}
-
-// Union represents Union plan.
-type Union struct {
-	basePlan
-
-	Selects []Plan
-}
-
 // Distinct represents Distinct plan.
 type Distinct struct {
 	baseLogicalPlan
-}
-
-// SetLimit implements Plan SetLimit interface.
-func (p *Distinct) SetLimit(limit float64) {
-	p.limit = limit
-	if p.GetChildByIndex(0) != nil {
-		p.GetChildByIndex(0).SetLimit(limit)
-	}
 }
 
 // Prepare represents prepare plan.
@@ -297,56 +137,6 @@ type Deallocate struct {
 	basePlan
 
 	Name string
-}
-
-// Aggregate represents a select fields plan.
-type Aggregate struct {
-	basePlan
-	AggFuncs     []*ast.AggregateFuncExpr
-	GroupByItems []*ast.ByItem
-}
-
-// SetLimit implements Plan SetLimit interface.
-func (p *Aggregate) SetLimit(limit float64) {
-	p.limit = limit
-	if p.GetChildByIndex(0) != nil {
-		p.GetChildByIndex(0).SetLimit(limit)
-	}
-}
-
-// Having represents a having plan.
-// The having plan should after aggregate plan.
-type Having struct {
-	basePlan
-
-	// Originally the WHERE or ON condition is parsed into a single expression,
-	// but after we converted to CNF(Conjunctive normal form), it can be
-	// split into a list of AND conditions.
-	Conditions []ast.ExprNode
-}
-
-// SetLimit implements Plan SetLimit interface.
-func (p *Having) SetLimit(limit float64) {
-	p.limit = limit
-	// We assume 50% of the GetChildByIndex(0) row is filtered out.
-	p.GetChildByIndex(0).SetLimit(limit * 2)
-}
-
-// Update represents an update plan.
-type Update struct {
-	basePlan
-
-	OrderedList []*ast.Assignment // OrderedList has the same offset as TablePlan's result fields.
-	SelectPlan  Plan
-}
-
-// Delete represents a delete plan.
-type Delete struct {
-	basePlan
-
-	SelectPlan   Plan
-	Tables       []*ast.TableName
-	IsMultiTable bool
 }
 
 // Filter represents a plan that filter GetChildByIndex(0)plan result.
